trabalhando com BigDecimal

+++ src/main/kotlin/br/com/pedro/allgames/extensions/InfoJogoExtension.kt
@@ -3,5 +3,5 @@ package br.com.pedro.allgames.extensions
 import br.com.pedro.allgames.model.InfoJogoJson
 import br.com.pedro.allgames.model.Jogo
 fun InfoJogoJson.criaJogo(): Jogo {
-    return Jogo(this.titulo, this.capa, this.preco, this.descricao)
+    return Jogo(this.titulo, this.capa, this.preco.toBigDecimal(), this.descricao)
 }
